Working on the entity/object system. Work in progress.
Not quite sure the best way to store world objects considering they can be different classes, have different attributes and can be pulled and used in different ways. Or how to reference them in multiple places.
Likely use a unique ID and an object manager. Not sure about casting types or if that is even allowed in Rust.
May use traits or an enum for the different object/entity types.